packages/browseros/pyproject.toml
[project]
name = "browseros"
version = "0.29.0"
description = "BrowserOS Build System"
requires-python = ">=3.12"
dependencies = [
  "click>=8.0.0",
  "typer>=0.20.0",
  "PyYAML>=5.4.1",
  "requests>=2.25.1",
  "google-cloud-storage>=2.10.0",
  "rich>=10.11.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
browseros-cli = "cli.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["cli", "build"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | env
  | _build
  | buck-out
  | build
  | dist
  | chromium_src
  | chromium_src_bak
  | third_party
)/
'''

[tool.isort]
profile = "black"
line_length = 88
skip_glob = ["env/*", "chromium_src/*", "chromium_src_bak/*", "third_party/*"]
